requirements.txt
appengine-python-standard ~= 1.1
devtools ~= 0.11
protobuf ~= 4.23
ruamel.yaml ~= 0.17
six ~= 1.16
urllib3 ~= 1.26
google-crc32c ~= 1.5
